Add /admin to finagle server

+++ thrift-swift-finagle-example-util/src/main/java/com/example/config/TracingConfig.java
@@ -6,5 +6,5 @@ public class TracingConfig {
     public String server;
     public Integer port;
     @JsonProperty("sample-rate")
-    public Integer sampleRate;
+    public Float sampleRate;
 }
